test: refactor test cases
